+++ gcc/SimpleDPP.h
@@ -50,7 +50,7 @@ __attribute__((weak)) byte __recv_data[SIMPLEDDP_DEFAULT_BUFFER_SIZE];
 void SimpleDPP_init(int send_capacity, int recv_capacity);
 void SimpleDPP_default_init();
 int SimpleDPP_send(const byte *data, int len);
-int __SimpleDPP_send_datas(const byte *data,size_t data_len,...);
+int __SimpleDPP_send_datas(const byte *data,int data_len,...);
 
 
 // Only works in C99
